test(runtime): retire spent dev-assist tests, split identity tests by kind (#2531)
The WP10 test file test_bridge_identity.py had grown into a 5-section
mega-file mixing genuine regression guards with development-assist
scaffolding. Retired the scaffolding and split the survivors by concern.

Deleted (outlived their purpose):
- NFR-006 timing-parity test — a one-shot pre/post-extraction perf gate that
  git-show+exec'd a pre-extraction snapshot; degrades to a self-compare
  waiver once the extraction is history.
- NFR-005 residual-LOC recorder — only asserted the file shrank at all.
- Two shape-only compat checks (seam-defines-symbol, native-delegate
  __module__) already covered family-wide by test_bridge_compat_surface.py.

Split the rest by kind:
- test_runtime_identity_resolution.py — functional resolution behavior
  (coord-branch + mission-ulid, incl. the fail-closed BranchIdentityUnresolved
  guard for the #2091-class git-worktree exit-128 scar).
- test_runtime_identity_seam_wiring.py — behavioral compat: delegate
  forwarding + the live-lookup regressions that keep the sibling
  test_runtime_bridge_identity.py monkeypatches effective.
- test_runtime_bridge_family_arch.py — standing arch invariants: zero-C901
  and no-import-cycle/DAG across the runtime_bridge* family.

Updated the now-stale test-file pointer in runtime_bridge_identity.py's
docstring. 15 tests (6+3+6) green.
